Move lockfile models into v1 directory
